Run only one iteration and move test to another file
